+++ client/CHANGELOG.md
@@ -1,8 +1,9 @@
 ## HEAD
 
-## v1.0.0-alpha.0
+- ⚙️ [v1.1.0] `pages/MenuBar`: Make the menu more simple to reflect the new way to access to exercises (#59).
+- ⚙️ [v1.1.0] `pages/links`: Add the exercise name to the link's title when playing (#56).
 
-### 💅 Components
+## v1.0.0-alpha.0
 
 - 🆕 [v1.0.0] `FrontPage`.
 - 🆕 [v1.0.0] `pages/MenuBar`.